+++ sso/user/tests/test_admin.py
@@ -1,25 +1,33 @@
 from collections import OrderedDict
+from datetime import timedelta
 from unittest import TestCase
 from unittest.mock import patch
 
 from django.test import Client
-from django.core.urlresolvers import reverse
+from django.contrib.admin import site
+from django.urls import reverse
+from django.conf import settings
+from django.utils import timezone
 
+from freezegun import freeze_time
 import pytest
 from allauth.account.models import EmailAddress
 
+from sso.user import admin
 from sso.user.models import User
 from sso.user.tests.factories import UserFactory
 from sso.oauth2.tests.factories import AccessTokenFactory, ApplicationFactory
+
+from core.tests.test_helpers import reload_urlconf
 
 
 @pytest.mark.django_db
 class DownloadCaseStudyCSVTestCase(TestCase):
 
     header = (
-        'created,date_joined,email,hashed_uuid,'
+        'created,date_joined,email,first_name,hashed_uuid,'
         'id,is_active,is_staff,is_superuser,'
-        'last_login,modified,oauth2_provider_application,user_profile,'
+        'last_login,last_name,modified,oauth2_provider_application,user_profile,'
         'utm,verification_code'
     )
 
@@ -46,9 +54,9 @@ class DownloadCaseStudyCSVTestCase(TestCase):
 
         user = User.objects.first()
         row_one = (
-            "{created},{date_joined},admin@example.com,{hashed_uuid},"
+            "{created},{date_joined},admin@example.com,,{hashed_uuid},"
             "{id},True,True,True,"
-            "{last_login},{modified},,,{utm},"
+            "{last_login},,{modified},,,{utm},"
         ).format(
             created=user.created,
             date_joined=user.date_joined,
@@ -82,9 +90,9 @@ class DownloadCaseStudyCSVTestCase(TestCase):
 
         user_one = User.objects.all()[2]
         row_one = (
-            '{created},{date_joined},{email},{hashed_uuid},'
+            '{created},{date_joined},{email},,{hashed_uuid},'
             '{id},{is_active},{is_staff},'
-            '{is_superuser},,{modified},,{user_profile},''{utm},'
+            '{is_superuser},,,{modified},,{user_profile},''{utm},'
             '{verification_code}'
         ).format(
             created=user_one.created,
@@ -103,9 +111,9 @@ class DownloadCaseStudyCSVTestCase(TestCase):
 
         user_two = User.objects.all()[1]
         row_two = (
-            '{created},{date_joined},{email},{hashed_uuid},'
+            '{created},{date_joined},{email},,{hashed_uuid},'
             '{id},{is_active},{is_staff},'
-            '{is_superuser},,{modified},,{user_profile},{utm},'
+            '{is_superuser},,,{modified},,{user_profile},{utm},'
             '{verification_code}'
         ).format(
             created=user_two.created,
@@ -124,9 +132,9 @@ class DownloadCaseStudyCSVTestCase(TestCase):
 
         user_three = User.objects.all()[0]
         row_three = (
-            '{created},{date_joined},{email},{hashed_uuid},'
+            '{created},{date_joined},{email},,{hashed_uuid},'
             '{id},{is_active},{is_staff},'
-            '{is_superuser},{last_login},{modified},,{user_profile},{utm},'
+            '{is_superuser},{last_login},,{modified},,{user_profile},{utm},'
             '{verification_code}'
         ).format(
             created=user_three.created,
@@ -154,9 +162,7 @@ class DownloadCaseStudyCSVTestCase(TestCase):
 
 @pytest.fixture
 def superuser():
-    return User.objects.create_superuser(
-        email='admin@example.com', password='test'
-    )
+    return User.objects.create_superuser(email='admin@example.com', password='test')
 
 
 @pytest.fixture
@@ -199,12 +205,14 @@ def test_download_csv_exops_not_fab(
         ('created', user_one.created),
         ('date_joined', user_one.date_joined),
         ('email', user_one.email),
+        ('first_name', ''),
         ('hashed_uuid', user_one.hashed_uuid),
         ('id', user_one.id),
         ('is_active', user_one.is_active),
         ('is_staff', user_one.is_staff),
         ('is_superuser', user_one.is_superuser),
         ('last_login', user_one.last_login),
+        ('last_name', ''),
         ('modified', user_one.modified),
         ('oauth2_provider_application', ''),
         ('user_profile', ''),
@@ -213,7 +221,6 @@ def test_download_csv_exops_not_fab(
     ])
 
     actual = str(response.content, 'utf-8').split('\r\n')
-
     assert actual[0] == ','.join(expected_row.keys())
     assert actual[1] == ','.join(map(str, expected_row.values()))
 
@@ -291,3 +298,68 @@ def test_download_email_verification_links(settings, superuser_client):
     )
 
     assert '/accounts/confirm-email/' in str(response.content)
+
+
+@pytest.mark.django_db
+class CompanyAdminAuthTestCase(TestCase):
+
+    def setUp(self):
+        self.client = Client()
+
+    AUTHENTICATION_BACKENDS_CLASSES = (
+        'authbroker_client.backends.AuthbrokerBackend',
+        'oauth2_provider.backends.OAuth2Backend',
+        'django.contrib.auth.backends.ModelBackend',
+        'allauth.account.auth_backends.AuthenticationBackend'
+    )
+
+    @pytest.mark.django_db
+    def test_nonsuperuser_cannot_access_user_changelist(self):
+        non_admin_user = UserFactory(is_staff=False)
+        non_admin_user.save()
+
+        settings.AUTHENTICATION_BACKENDS = self.AUTHENTICATION_BACKENDS_CLASSES
+        settings.FEATURE_ENFORCE_STAFF_SSO_ENABLED = True
+        reload_urlconf()
+        self.client.force_login(non_admin_user)
+        url = reverse('admin:user_user_changelist')
+
+        response = self.client.get(url)
+        assert response.status_code == 302
+
+    @pytest.mark.django_db
+    def test_no_user_cannot_access_user_changelist(self):
+
+        settings.AUTHENTICATION_BACKENDS = self.AUTHENTICATION_BACKENDS_CLASSES
+        settings.FEATURE_ENFORCE_STAFF_SSO_ENABLED = True
+        reload_urlconf()
+
+        url = reverse('admin:user_user_changelist')
+
+        response = self.client.get(url)
+        assert response.status_code == 302
+
+
+@pytest.mark.django_db
+def test_GDPR_compliance_filter(rf, superuser):
+    three_years_ago = 365 * 3
+
+    with freeze_time(timezone.now() - timedelta(days=three_years_ago + 1)):
+        user_one = UserFactory()
+
+    with freeze_time(timezone.now() - timedelta(days=three_years_ago)):
+        user_two = UserFactory()
+
+    with freeze_time(timezone.now() - timedelta(days=three_years_ago - 1)):
+        user_three = UserFactory()
+
+    modeladmin = admin.UserAdmin(User, site)
+    request = rf.get('/', {'gdpr': True})
+    request.user = superuser
+    changelist = modeladmin.get_changelist_instance(request)
+    queryset = changelist.get_queryset(request)
+
+    assert queryset.count() == 2
+    assert user_one in queryset
+    assert user_two in queryset
+    assert user_three not in queryset